% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_classification.R
\name{confusion_matrix}
\alias{confusion_matrix}
\title{Binary confusion matrix}
\usage{
confusion_matrix(actual, predicted)
}
\arguments{
\item{actual}{The ground truth binary numeric vector containing 1 for the positive
class and 0 for the negative class.}

\item{predicted}{The predicted binary numeric vector containing 1 for the positive
class and 0 for the negative class. Each element represents the
prediction for the corresponding element in \code{actual}.}
}
\description{
\code{confusion_matrix} Calculates a binary classification confusion matrix,
comparing the predicted with the actual values for the classes.
Assumes that 1 is used for the positive class and 0 for the
negative class.
}
\details{
Returns a \code{data.frame} with columns corresponding to the
number of True Positives (\code{tp}), False Positives (\code{fp}),
True Negatives (\code{tn}), and False Negatives (\code{fn})
}
\examples{
actual <- c(1, 1, 1, 0, 0, 0)
predicted <- c(1, 0, 1, 1, 1, 1)
confusion_matrix(actual, predicted)
}
\seealso{
\code{\link{sensitivity}} \code{\link{specificity}}
}
