% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_classification.R
\name{fdr}
\alias{fdr}
\title{False Discovery Rate}
\usage{
fdr(actual, predicted)
}
\arguments{
\item{actual}{The ground truth binary numeric vector containing 1 for the positive
class and 0 for the negative class.}

\item{predicted}{The predicted binary numeric vector containing 1 for the positive
class and 0 for the negative class. Each element represents the
prediction for the corresponding element in \code{actual}.}
}
\description{
\code{fdr} computes proportion of observations predicted to be in
the positive class (i.e. the element in \code{predicted} equals 1) that
actually belong to the negative class (i.e.the element in \code{actual} equals 0).
}
\details{
It is implemented as \code{1 - ppv}
}
\examples{
actual <- c(1, 1, 1, 0, 0, 0)
predicted <- c(1, 0, 1, 1, 1, 1)
fpr(actual, predicted)
}
\seealso{
\code{\link{confusion_matrix}} \code{\link{ppv}}
}
