% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_classification.R
\name{fnr}
\alias{fnr}
\title{False Negative Rate}
\usage{
fnr(actual, predicted)
}
\arguments{
\item{actual}{The ground truth binary numeric vector containing 1 for the positive
class and 0 for the negative class.}

\item{predicted}{The predicted binary numeric vector containing 1 for the positive
class and 0 for the negative class. Each element represents the
prediction for the corresponding element in \code{actual}.}
}
\description{
\code{fnr} calculates the proportion of actual positives (\code{actual} equals 1)
that are not identified as such.
}
\details{
It is defined as \code{1 - sensitivity}
}
\examples{
actual <- c(1, 1, 1, 0, 0, 0)
predicted <- c(1, 0, 1, 1, 1, 1)
fnr(actual, predicted)
}
\seealso{
\code{\link{confusion_matrix}} \code{\link{sensitivity}}
}
