% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_classification.R
\name{ll}
\alias{ll}
\title{Log Loss}
\usage{
ll(actual, predicted, eps = 1e-15)
}
\arguments{
\item{actual}{The ground truth binary numeric vector containing 1 for the positive
class and 0 for the negative class.}

\item{predicted}{A numeric vector of predicted values, where the values correspond
to the probabilities that each observation in \code{actual}
belongs to the positive class}

\item{eps}{Log loss is undefined for p=0 or p=1, so probabilities are clipped to 
\code{pmax(eps, pmin(1 - eps, p))}.}
}
\description{
\code{ll} computes the elementwise log loss between two numeric vectors.
}
\examples{
actual <- c(1, 1, 1, 0, 0, 0)
predicted <- c(0.9, 0.8, 0.4, 0.5, 0.3, 0.2)
ll(actual, predicted)
}
\seealso{
\code{\link{logLoss}}
}
