% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_classification.R
\name{lrn}
\alias{lrn}
\title{Negative Likelihood Ratio}
\usage{
lrn(actual, predicted)
}
\arguments{
\item{actual}{The ground truth binary numeric vector containing 1 for the positive
class and 0 for the negative class.}

\item{predicted}{The predicted binary numeric vector containing 1 for the positive
class and 0 for the negative class. Each element represents the
prediction for the corresponding element in \code{actual}.}
}
\description{
\code{lrn} is used to assessing the value of performing a
diagnostic test, and estimates the ratio of the probability
of a true negative result over the probability of a false negative
result.
}
\details{
It is implemented as the ratio: \code{(1 - sensitivity) / specificity}
}
\examples{
actual <- c(1, 1, 1, 0, 0, 0)
predicted <- c(1, 0, 1, 1, 1, 1)
lrn(actual, predicted)
}
\seealso{
\code{\link{specificity}} \code{\link{sensitivity}}
}
