% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_classification.R
\name{lrp}
\alias{lrp}
\title{Positive Likelihood Ratio}
\usage{
lrp(actual, predicted)
}
\arguments{
\item{actual}{The ground truth binary numeric vector containing 1 for the positive
class and 0 for the negative class.}

\item{predicted}{The predicted binary numeric vector containing 1 for the positive
class and 0 for the negative class. Each element represents the
prediction for the corresponding element in \code{actual}.}
}
\description{
\code{lrp} is used to assessing the value of performing a
diagnostic test, and estimates the ratio of the probability
of a true positive result over the probability of a false positive
result.
}
\details{
It is implemented as the ratio: \code{sensitivity / (1 - specificity)}
}
\examples{
actual <- c(1, 1, 1, 0, 0, 0)
predicted <- c(1, 0, 1, 1, 1, 1)
lrp(actual, predicted)
}
\seealso{
\code{\link{tpr}} \code{\link{fpr}}
}
