% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_classification.R
\name{ppv}
\alias{ppv}
\title{Positive Predictive Value}
\usage{
ppv(actual, predicted)
}
\arguments{
\item{actual}{The ground truth binary numeric vector containing 1 for the positive
class and 0 for the negative class.}

\item{predicted}{The predicted binary numeric vector containing 1 for the positive
class and 0 for the negative class. Each element represents the
prediction for the corresponding element in \code{actual}.}
}
\description{
\code{ppv} calculates the proportion of all predicted positive values (\code{predicted} equals 1) that
are true positive results. It is also known as \code{precision}
}
\examples{
actual <- c(1, 1, 1, 0, 0, 0)
predicted <- c(1, 0, 1, 1, 1, 1)
ppv(actual, predicted)
}
\seealso{
\code{\link{confusion_matrix}} \code{\link{npv}} \code{\link{precision}}
}
