% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_classification.R
\name{specificity}
\alias{specificity}
\title{Specificity}
\usage{
specificity(actual, predicted)
}
\arguments{
\item{actual}{The ground truth binary numeric vector containing 1 for the positive
class and 0 for the negative class.}

\item{predicted}{The predicted binary numeric vector containing 1 for the positive
class and 0 for the negative class. Each element represents the
prediction for the corresponding element in \code{actual}.}
}
\description{
\code{specificity} calculates the proportion of actual negatives (\code{actual} equals 0)
that are correctly identified as such. It is also known as
\code{true negative rate}.
}
\examples{
actual <- c(1, 1, 1, 0, 0, 0)
predicted <- c(1, 0, 1, 1, 1, 1)
specificity(actual, predicted)
}
\seealso{
\code{\link{confusion_matrix}} \code{\link{sensitivity}}
}
